feat: selecting and resetting map
